chore: rename to kebab-case
