+++ jest.config.js
@@ -18,11 +18,11 @@ module.exports = {
   },
   testRegex: ["/test/.*\\.(test.js|test.ts)$"],
   moduleFileExtensions: ["ts", "js", "json"],
-  snapshotResolver: "<rootDir>/scripts/snapshotResolver.js",
+  snapshotResolver: "<rootDir>/scripts/snapshot-resolver.js",
   watchPlugins: ["jest-watch-typeahead/filename", "jest-watch-typeahead/testname"],
-  setupFilesAfterEnv: ["<rootDir>/scripts/setupTest.js"],
-  globalTeardown: "<rootDir>/scripts/cleanupTest.js",
-  globalSetup: "<rootDir>/scripts/globalSetup.js",
+  setupFilesAfterEnv: ["<rootDir>/scripts/setup-test.js"],
+  globalTeardown: "<rootDir>/scripts/cleanup-test.js",
+  globalSetup: "<rootDir>/scripts/global-setup.js",
   modulePathIgnorePatterns: [
     "<rootDir>/test/loader/test-loader",
     "<rootDir>/test/plugin/test-plugin",
