+++ scripts/cleanup-test.js
