+++ scripts/global-setup.js
