+++ scripts/setup-build.js
