+++ scripts/snapshot-resolver.js
