+++ test/api/resolve-config/resolve-config.test.js
