+++ test/api/resolve-config/webpack.config2.cjs
